Subtotal e total do servico

+++ src/main/java/com/mpinfo/servicosprof/repositories/ChamadoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.mpinfo.servicosprof.domain.Chamado;
 
 @Repository
-public interface ChamadoRepository extends JpaRepository<Chamado, Long> {
+public interface ChamadoRepository extends JpaRepository<Chamado, Integer> {
 
 }
